+++ evosax/strategies/ma_es.py
@@ -3,19 +3,18 @@ import jax.numpy as jnp
 import chex
 from typing import Tuple
 from ..strategy import Strategy
-from ..utils.eigen_decomp import full_eigen_decomp
 
 
-class CMA_ES(Strategy):
+class MA_ES(Strategy):
     def __init__(self, num_dims: int, popsize: int, elite_ratio: float = 0.5):
-        """CMA-ES (e.g. Hansen, 2016)
-        Reference: https://arxiv.org/abs/1604.00772
-        Inspired by: https://github.com/CyberAgentAILab/cmaes"""
+        """MA-ES (Bayer & Sendhoff, 2017)
+        Reference: https://www.honda-ri.de/pubs/pdf/3376.pdf
+        """
         super().__init__(num_dims, popsize)
         assert 0 <= elite_ratio <= 1
         self.elite_ratio = elite_ratio
         self.elite_popsize = int(self.popsize * self.elite_ratio)
-        self.strategy_name = "CMA_ES"
+        self.strategy_name = "MA_ES"
 
     @property
     def params_strategy(self) -> chex.ArrayTree:
@@ -103,12 +102,9 @@ class CMA_ES(Strategy):
         )
         state = {
             "p_sigma": jnp.zeros(self.num_dims),
-            "p_c": jnp.zeros(self.num_dims),
             "sigma": params["sigma_init"],
             "mean": initialization,
-            "C": jnp.eye(self.num_dims),
-            "D": None,
-            "B": None,
+            "M": jnp.eye(self.num_dims),
         }
         return state
 
@@ -116,17 +112,14 @@ class CMA_ES(Strategy):
         self, rng: chex.PRNGKey, state: chex.ArrayTree, params: chex.ArrayTree
     ) -> Tuple[chex.Array, chex.ArrayTree]:
         """`ask` for new parameter candidates to evaluate next."""
-        C, B, D = full_eigen_decomp(state["C"], state["B"], state["D"])
         x = sample(
             rng,
             state["mean"],
             state["sigma"],
-            B,
-            D,
+            state["M"],
             self.num_dims,
             self.popsize,
         )
-        state["C"], state["B"], state["D"] = C, B, D
         return x, state
 
     def tell_strategy(
@@ -141,27 +134,17 @@ class CMA_ES(Strategy):
         concat_p_f = jnp.hstack([jnp.expand_dims(fitness, 1), x])
         sorted_solutions = concat_p_f[concat_p_f[:, 0].argsort()]
         # Update mean, isotropic/anisotropic paths, covariance, stepsize
-        y_k, y_w, mean = update_mean(
+        mean, z_k = update_mean(
             state["mean"], state["sigma"], sorted_solutions, params
         )
 
-        p_sigma, C_2, C, B, D = update_p_sigma(
-            state["C"], state["B"], state["D"], state["p_sigma"], y_w, params
-        )
-
-        p_c, norm_p_sigma, h_sigma = update_p_c(
-            mean, p_sigma, state["p_c"], state["gen_counter"] + 1, y_w, params
-        )
-
-        C = update_covariance(mean, p_c, C, y_k, h_sigma, C_2, params)
+        p_sigma, norm_p_sigma = update_p_sigma(z_k, state["p_sigma"], params)
+        M = update_M_matrix(state["M"], state["p_sigma"], z_k, params)
         sigma = update_sigma(state["sigma"], norm_p_sigma, params)
 
         state["mean"] = mean
         state["p_sigma"] = p_sigma
-        state["C"] = C
-        state["B"] = B
-        state["D"] = D
-        state["p_c"] = p_c
+        state["M"] = M
         state["sigma"] = sigma
         return state
 
@@ -171,86 +154,49 @@ def update_mean(
     sigma: float,
     sorted_solutions: chex.Array,
     params: chex.ArrayTree,
-) -> Tuple[chex.Array, chex.Array, chex.Array]:
+) -> Tuple[chex.Array, chex.Array]:
     """Update mean of strategy."""
-    x_k = sorted_solutions[:, 1:]  # ~ N(m, σ^2 C)
-    y_k = (x_k - mean) / sigma  # ~ N(0, C)
+    z_k = sorted_solutions[:, 1:] - mean  # ~ N(0, σ^2 C)
+    y_k = z_k / sigma  # ~ N(0, C)
     y_w = jnp.sum(y_k.T * params["weights_truncated"], axis=1)
     mean += params["c_m"] * sigma * y_w
-    return y_k, y_w, mean
+    return mean, z_k
 
 
 def update_p_sigma(
-    C: chex.Array,
-    B: chex.Array,
-    D: chex.Array,
+    z_k: chex.Array,
     p_sigma: chex.Array,
-    y_w: chex.Array,
     params: chex.ArrayTree,
-) -> Tuple[chex.Array, chex.Array, chex.Array, None, None]:
+) -> Tuple[chex.Array, float]:
     """Update evolution path for covariance matrix."""
-    C, B, D = full_eigen_decomp(C, B, D)
-    C_2 = B.dot(jnp.diag(1 / D)).dot(B.T)  # C^(-1/2) = B D^(-1) B^T
+    z_w = jnp.sum(z_k.T * params["weights_truncated"], axis=1)
     p_sigma_new = (1 - params["c_sigma"]) * p_sigma + jnp.sqrt(
         params["c_sigma"] * (2 - params["c_sigma"]) * params["mu_eff"]
-    ) * C_2.dot(y_w)
-    _B, _D = None, None
-    return p_sigma_new, C_2, C, _B, _D
+    ) * z_w
+    norm_p_sigma = jnp.linalg.norm(p_sigma_new)
+    return p_sigma_new, norm_p_sigma
 
 
-def update_p_c(
-    mean: chex.Array,
-    p_sigma: chex.Array,
-    p_c: chex.Array,
-    gen_counter: int,
-    y_w: chex.Array,
-    params: chex.ArrayTree,
-) -> Tuple[chex.Array, float, float]:
-    """Update evolution path for sigma/stepsize."""
-    norm_p_sigma = jnp.linalg.norm(p_sigma)
-    h_sigma_cond_left = norm_p_sigma / jnp.sqrt(
-        1 - (1 - params["c_sigma"]) ** (2 * (gen_counter))
-    )
-    h_sigma_cond_right = (1.4 + 2 / (mean.shape[0] + 1)) * params["chi_n"]
-    h_sigma = 1.0 * (h_sigma_cond_left < h_sigma_cond_right)
-    p_c_new = (1 - params["c_c"]) * p_c + h_sigma * jnp.sqrt(
-        params["c_c"] * (2 - params["c_c"]) * params["mu_eff"]
-    ) * y_w
-    return p_c_new, norm_p_sigma, h_sigma
-
-
-def update_covariance(
-    mean: chex.Array,
-    p_c: chex.Array,
-    C: chex.Array,
-    y_k: chex.Array,
-    h_sigma: float,
-    C_2: chex.Array,
-    params: chex.ArrayTree,
+def update_M_matrix(
+    M: chex.Array, p_sigma: chex.Array, z_k: chex.Array, params: chex.ArrayTree
 ) -> chex.Array:
-    """Update cov. matrix estimator using rank 1 + μ updates."""
-    w_io = params["weights"] * jnp.where(
-        params["weights"] >= 0,
-        1,
-        mean.shape[0] / (jnp.linalg.norm(C_2.dot(y_k.T), axis=0) ** 2 + 1e-20),
-    )
-    delta_h_sigma = (1 - h_sigma) * params["c_c"] * (2 - params["c_c"])
-    rank_one = jnp.outer(p_c, p_c)
+    """Update the M matrix."""
+    rank_one = jnp.outer(p_sigma, p_sigma)
     rank_mu = jnp.sum(
-        jnp.array([w * jnp.outer(y, y) for w, y in zip(w_io, y_k)]), axis=0
+        jnp.array(
+            [
+                w * jnp.outer(z, z)
+                for w, z in zip(params["weights_truncated"], z_k)
+            ]
+        ),
+        axis=0,
     )
-    C = (
-        (
-            1
-            + params["c_1"] * delta_h_sigma
-            - params["c_1"]
-            - params["c_mu"] * jnp.sum(params["weights"])
-        )
-        * C
-        + params["c_1"] * rank_one
-        + params["c_mu"] * rank_mu
+    M_new = M @ (
+        jnp.eye(M.shape[0])
+        + params["c_1"] / 2 * (rank_one - jnp.eye(M.shape[0]))
+        + params["c_mu"] / 2 * (rank_mu - jnp.eye(M.shape[0]))
     )
-    return C
+    return M_new
 
 
 def update_sigma(
@@ -268,14 +214,13 @@ def sample(
     rng: chex.PRNGKey,
     mean: chex.Array,
     sigma: float,
-    B: chex.Array,
-    D: chex.Array,
+    M: chex.Array,
     n_dim: int,
     pop_size: int,
 ) -> chex.Array:
     """Jittable Gaussian Sample Helper."""
     z = jax.random.normal(rng, (n_dim, pop_size))  # ~ N(0, I)
-    y = B.dot(jnp.diag(D)).dot(z)  # ~ N(0, C)
+    y = M.dot(z)  # ~ N(0, C)
     y = jnp.swapaxes(y, 1, 0)
     x = mean + sigma * y  # ~ N(m, σ^2 C)
     return x